+++ src/modules/app/app.module.ts
@@ -3,7 +3,6 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
 
 import config from '../../config';
-import { HairServiceModule } from '../hair-service/hair-service.module';
 import { SheetModule } from '../sheet/sheet.module';
 import { ClientModule } from '../client/client.module';
 import { ColoristModule } from '../colorist/colorist.module';
@@ -24,7 +23,6 @@ import { AuthModule } from '../auth/auth.module';
     }),
     ClientModule,
     ColoristModule,
-    HairServiceModule,
     SheetModule,
   ],
 })
