+++ src/modules/client/client.module.ts
@@ -6,14 +6,12 @@ import { ClientService } from './client.service';
 import { Client, ClientSchema } from './schemas';
 import { ColoristModule } from '../colorist/colorist.module';
 import { SheetModule } from '../sheet/sheet.module';
-import { HairServiceModule } from '../hair-service/hair-service.module';
 
 @Module({
   controllers: [ClientController],
   exports: [ClientService],
   imports: [
     forwardRef(() => ColoristModule),
-    forwardRef(() => HairServiceModule),
     forwardRef(() => SheetModule),
     MongooseModule.forFeature([{ name: Client.name, schema: ClientSchema }]),
   ],
