+++ src/modules/colorist/colorist.module.ts
@@ -6,14 +6,12 @@ import { ColoristController } from './colorist.controller';
 import { ColoristService } from './colorist.service';
 import { ClientModule } from '../client/client.module';
 import { SheetModule } from '../sheet/sheet.module';
-import { HairServiceModule } from '../hair-service/hair-service.module';
 
 @Module({
   controllers: [ColoristController],
   exports: [ColoristService],
   imports: [
     forwardRef(() => ClientModule),
-    forwardRef(() => HairServiceModule),
     forwardRef(() => SheetModule),
     MongooseModule.forFeature([
       { name: Colorist.name, schema: ColoristSchema },
