--- src/modules/hair-service/constants/hair-service-ingredient.constant.ts
@@ -1,27 +1,0 @@
-export enum HAIR_SERVICE_INGREDIENT_BRAND_LENGTH {
-  MAX = 50,
-  MIN = 2,
-}
-
-export enum HAIR_SERVICE_INGREDIENT_HEIGHT {
-  MAX = 10,
-  MIN = 0,
-}
-
-export enum HAIR_SERVICE_INGREDIENT_OXIDIZING {
-  MAX = 100,
-  MIN = 0,
-}
-
-export enum HAIR_SERVICE_INGREDIENT_QUANTITY_LENGTH {
-  MAX = 10,
-  MIN = 1,
-}
-
-export enum HAIR_SERVICE_INGREDIENT_TONE_LENGTH {
-  MAX = 30,
-  MIN = 1,
-}
-
-export const HAIR_SERVICE_INGREDIENT_HEIGHT_MAX_DECIMAL_PLACES = 0;
-export const HAIR_SERVICE_INGREDIENT_OXIDIZING_MAX_DECIMAL_PLACES = 2;
