--- src/modules/hair-service/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './hair-service-ingredient.constant';
-export * from './hair-service.constant';
