--- src/modules/hair-service/dtos/index.ts
@@ -1,2 +1,0 @@
-export * from './hair-service-ingredient.dto';
-export * from './hair-service.dto';
