--- src/modules/hair-service/interfaces/hair-service-ingredient.interface.ts
@@ -1,9 +1,0 @@
-export interface IHairServiceIngredient {
-  brand: string;
-  height: number;
-  oxidizing: number;
-  quantity: string;
-  tone: string;
-}
-
-export type ICreateHairServiceIngredient = IHairServiceIngredient;
