--- src/modules/hair-service/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './hair-service.interface';
-export * from './hair-service-ingredient.interface';
