--- src/modules/hair-service/schemas/index.ts
@@ -1,1 +1,0 @@
-export * from './hair-service.schema';
