+++ src/modules/sheet/constants/index.ts
@@ -1,1 +1,3 @@
 export * from './sheet.constant';
+export * from './hair-service-ingredient.constant';
+export * from './hair-service.constant';
