+++ src/modules/sheet/dtos/index.ts
@@ -1,1 +1,3 @@
 export * from './sheet.dto';
+export * from './hair-service-ingredient.dto';
+export * from './hair-service.dto';
