+++ src/modules/sheet/dtos/sheet.dto.ts
@@ -1,9 +1,4 @@
-import {
-  ApiProperty,
-  IntersectionType,
-  OmitType,
-  PickType,
-} from '@nestjs/swagger';
+import { ApiProperty, IntersectionType, PartialType } from '@nestjs/swagger';
 import { ArrayMaxSize, IsMongoId } from 'class-validator';
 
 import {
@@ -15,29 +10,23 @@ import {
 
 import {
   IChangeClientDto,
-  ICreateHairServiceInSheet,
   ICreateSheetDto,
-  ICreateSheetResponseDto,
   IFindSheetsQueryDto,
-  ISheetDto,
+  IHairService,
+  ISheet,
 } from '../interfaces';
 
-import { IHairServiceDto } from '../../hair-service/interfaces';
-import { CreateHairServiceDto, HairServiceDto } from '../../hair-service/dtos';
 import { IsSheetDate } from '../decorators';
 import { SHEET_MAX_HAIR_SERVICES } from '../constants';
-
-class CreateHairServiceInSheetDto
-  extends OmitType(CreateHairServiceDto, ['clientId', 'sheetId'])
-  implements ICreateHairServiceInSheet {}
+import { HairServiceDto } from './hair-service.dto';
 
 export class CreateSheetDto implements ICreateSheetDto {
   @ApiPropertyDto({
-    dto: CreateHairServiceInSheetDto,
+    dto: HairServiceDto,
     isArray: true,
   })
   @ArrayMaxSize(SHEET_MAX_HAIR_SERVICES)
-  hairServices: ICreateHairServiceInSheet[];
+  hairServices: IHairService[];
 
   @ApiProperty({
     description: `Client's _id to which this sheet belongs`,
@@ -55,28 +44,12 @@ export class CreateSheetDto implements ICreateSheetDto {
 }
 
 export class UpdateSheetDto
-  extends PickType(CreateSheetDto, ['date'])
-  implements Pick<ICreateSheetDto, 'date'> {}
+  extends PartialType(CreateSheetDto)
+  implements Partial<ICreateSheetDto> {}
 
 export class SheetDto
   extends IntersectionType(CreateSheetDto, BasicDocumentDto, ColoristIdDto)
-  implements ISheetDto
-{
-  @ApiPropertyDto({ dto: HairServiceDto, isArray: true })
-  hairServices: IHairServiceDto[];
-}
-
-export class CreateSheetResponseDto
-  extends IntersectionType(
-    OmitType(CreateSheetDto, ['hairServices']),
-    BasicDocumentDto,
-    ColoristIdDto,
-  )
-  implements ICreateSheetResponseDto
-{
-  @ApiPropertyDto({ dto: HairServiceDto, isArray: true })
-  hairServices: IHairServiceDto[];
-}
+  implements ISheet {}
 
 export class FindSheetsQueryDto
   extends BasicQueryDto
