+++ src/modules/sheet/interfaces/hair-service-ingredient.interface.ts
@@ -1,0 +1,7 @@
+export interface IHairServiceIngredient {
+  brand: string;
+  height?: string;
+  oxidizing: string;
+  quantity: string;
+  tone: string;
+}
