+++ src/modules/sheet/interfaces/hair-service.interface.ts
@@ -1,0 +1,7 @@
+import { IHairServiceIngredient } from './hair-service-ingredient.interface';
+
+export interface IHairService {
+  ingredients: IHairServiceIngredient[];
+  name: string;
+  observations?: string;
+}
