+++ src/modules/sheet/interfaces/index.ts
@@ -1,1 +1,3 @@
 export * from './sheet.interface';
+export * from './hair-service.interface';
+export * from './hair-service-ingredient.interface';
