+++ src/modules/sheet/interfaces/sheet.interface.ts
@@ -1,60 +1,21 @@
 import { IBasicDocument, IBasicQueryDto, IColoristId } from '../../../common';
 import { IClientId } from '../../client/interfaces';
-import {
-  ICreateHairServiceDto,
-  IHairService,
-  IHairServiceDto,
-} from '../../hair-service/interfaces';
+import { IHairService } from './hair-service.interface';
 
 export interface ISheetId {
   sheetId: string;
 }
 
-export interface ISheetObjectIdAttributes {
+export interface ISheetAttributes extends IColoristId, IClientId {
+  date: string;
   hairServices: IHairService[];
 }
 
-interface ISheetDtoObjectIdAttributes {
-  hairServices: IHairServiceDto[];
-}
-
-export type ICreateHairServiceInSheet = Omit<
-  ICreateHairServiceDto,
-  'clientId' | 'sheetId'
->;
-
-interface ICreateSheetObjectIdAttributes {
-  hairServices: ICreateHairServiceInSheet[];
-}
-
-export interface ISheetAttributes extends IColoristId, IClientId {
-  date: string;
-}
-
-export interface ISheet
-  extends ISheetAttributes,
-    ISheetObjectIdAttributes,
-    IBasicDocument {}
-
-export interface ISheetDto
-  extends ISheetAttributes,
-    ISheetDtoObjectIdAttributes,
-    IBasicDocument {}
+export interface ISheet extends ISheetAttributes, IBasicDocument {}
 
 export type ICreateSheet = ISheetAttributes;
 
-export interface ICreateSheetWithHairServices
-  extends ISheetAttributes,
-    ICreateSheetObjectIdAttributes {}
-
-export interface ICreateSheetDto
-  extends Omit<ICreateSheet, 'coloristId'>,
-    ICreateSheetObjectIdAttributes {}
-
-export interface ICreateSheetResponseDto
-  extends ISheetDto,
-    IBasicDocument,
-    IColoristId {}
+export type ICreateSheetDto = Omit<ICreateSheet, 'coloristId'>;
 
 export type IFindSheetsQueryDto = IBasicQueryDto;
